LayoutTests/webanimations/animation-interface-start-time-property.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check the behavior of the startTime property on Animation objects.");

debug("The animation.startTime property should be null by default.");
const animation = new Animation;
shouldBeNull("animation.startTime");

debug("");
debug("The animation.startTime property can be set to a number.");
animation.startTime = 2;
shouldBe("animation.startTime", "2");

debug("");
debug("The animation.effect property can be set back to null.");
animation.startTime = null;
shouldBeNull("animation.startTime");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
