LayoutTests/webanimations/animation-interface-start-time-property-expected.txt

Check the behavior of the startTime property on Animation objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The animation.startTime property should be null by default.
PASS animation.startTime is null

The animation.startTime property can be set to a number.
PASS animation.startTime is 2

The animation.effect property can be set back to null.
PASS animation.startTime is null

PASS successfullyParsed is true

TEST COMPLETE
